requirements.txt
Flask==3.1.2
gunicorn==23.0.0
numpy==1.26.4
Pillow==11.3.0
xgboost-cpu==3.2.0
ai-edge-litert==2.1.3
